+++ src/main/kotlin/com/mpi/kissing_company/controllers/InviteLinksController.kt
@@ -2,12 +2,15 @@ package com.mpi.kissing_company.controllers
 
 import com.mpi.kissing_company.entities.InviteLinks
 import com.mpi.kissing_company.repositories.InviteLinksRepository
+import org.springframework.http.HttpStatus
+import org.springframework.transaction.annotation.Transactional
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
+import org.springframework.web.server.ResponseStatusException
 
 
 @RestController
@@ -24,8 +27,12 @@ internal class InviteLinksController(private val repository: InviteLinksReposito
         return repository.findAll()
     }
 
-    @DeleteMapping("/invite_link/{inviteToken}")  // FIXME: Delete and Put not allowed
+    @DeleteMapping("/invite_link/{inviteToken}")
+    @Transactional
     fun delete(@PathVariable inviteToken: String){
+        if (repository.existsById(inviteToken) == false){
+            throw ResponseStatusException(HttpStatus.NOT_FOUND, "Invite token not found")
+        }
         repository.deleteByinviteToken(inviteToken)
     }
 }
